refactor(automation): split event trigger dialog into section components
