+++ apps/dashboard/src/components/automation/events/event-trigger-event-section.tsx
@@ -1,0 +1,47 @@
+import { useStore } from "@tanstack/react-form";
+import { EVENT_TYPE_ORDER } from "@/constants/event-triggers";
+import type { EventTriggerFormSectionProps } from "@/types/automation/event-trigger";
+import { EventTypeCard } from "./event-type-card";
+import { TriggerSwitchRow } from "./trigger-switch-row";
+
+export function EventTriggerEventSection({
+  form,
+}: EventTriggerFormSectionProps) {
+  const eventType = useStore(form.store, (s) => s.values.eventType);
+
+  return (
+    <section className="space-y-3">
+      <div className="space-y-1">
+        <h3 className="font-semibold text-base">Trigger event</h3>
+        <p className="text-muted-foreground text-sm">When should this fire?</p>
+      </div>
+      <form.Field name="eventType">
+        {(field) => (
+          <div className="grid gap-3 md:grid-cols-2">
+            {EVENT_TYPE_ORDER.map((type) => (
+              <EventTypeCard
+                eventType={type}
+                key={type}
+                onSelect={() => field.handleChange(type)}
+                selected={field.state.value === type}
+              />
+            ))}
+          </div>
+        )}
+      </form.Field>
+      {eventType === "release" && (
+        <form.Field name="includePreReleases">
+          {(field) => (
+            <TriggerSwitchRow
+              checked={field.state.value}
+              id={field.name}
+              label="Include pre-releases"
+              onCheckedChange={field.handleChange}
+              tooltip="When off, releases marked as pre-release on GitHub will not fire this trigger."
+            />
+          )}
+        </form.Field>
+      )}
+    </section>
+  );
+}
