+++ apps/dashboard/src/types/automation/event-trigger.ts
@@ -1,4 +1,6 @@
+import type { useEventTriggerForm } from "@/lib/hooks/use-event-trigger-form";
 import type { WebhookEventType } from "@/schemas/integrations";
+import type { BrandSettings } from "@/types/hooks/brand-analysis";
 import type { Trigger } from "@/types/triggers/triggers";
 
 export type { EventTriggerFormValues } from "@/schemas/automation/event-trigger-form";
@@ -24,4 +26,58 @@ export interface TriggerSwitchRowProps {
   tooltip: string;
   checked: boolean;
   onCheckedChange: (checked: boolean) => void;
+}
+
+export interface UseEventTriggerFormProps {
+  organizationId: string;
+  editTrigger?: Trigger;
+  open: boolean;
+  onSuccess?: (trigger: Trigger) => void;
+  onClose: () => void;
+}
+
+export type EventTriggerFormApi = ReturnType<
+  typeof useEventTriggerForm
+>["form"];
+
+export interface EventTriggerFormSectionProps {
+  form: EventTriggerFormApi;
+}
+
+export interface RepositoryOption {
+  value: string;
+  label: string;
+}
+
+export interface EventTriggerRepositoriesSectionProps {
+  form: EventTriggerFormApi;
+  isLoading: boolean;
+  options: RepositoryOption[];
+  onAddRepository: () => void;
+}
+
+export interface EventTriggerRulesSectionProps {
+  form: EventTriggerFormApi;
+  brandVoices: BrandSettings[];
+}
+
+export interface EventTriggerDialogFooterProps {
+  errorMessage: string | null;
+  isEditMode: boolean;
+  isPending: boolean;
+  onCancel: () => void;
+  repositoryCount: number;
+}
+
+export interface EventTriggerFooterStatusProps {
+  errorMessage: string | null;
+  repositoryCount: number;
+}
+
+export interface RepositoryConnectionDialogsProps {
+  githubIntegrationId?: string;
+  isEditMode: boolean;
+  onOpenChange: (open: boolean) => void;
+  open: boolean;
+  organizationId: string;
 }
